Move number and troop rendering utilities to core
