+++ src/client/graphics/layers/RadialMenuElements.ts
@@ -2,8 +2,8 @@ import { Config } from "../../../core/configuration/Config";
 import { AllPlayers, PlayerActions, UnitType } from "../../../core/game/Game";
 import { TileRef } from "../../../core/game/GameMap";
 import { GameView, PlayerView } from "../../../core/game/GameView";
-import { flattenedEmojiTable } from "../../../core/Util";
-import { renderNumber, translateText } from "../../Utils";
+import { flattenedEmojiTable, renderNumber } from "../../../core/Util";
+import { translateText } from "../../Utils";
 import { BuildItemDisplay, BuildMenu, flattenedBuildTable } from "./BuildMenu";
 import { ChatIntegration } from "./ChatIntegration";
 import { EmojiTable } from "./EmojiTable";
