+++ src/client/graphics/layers/TeamStats.ts
@@ -3,7 +3,8 @@ import { customElement, property } from "lit/decorators.js";
 import { EventBus } from "../../../core/EventBus";
 import { GameMode, Team } from "../../../core/game/Game";
 import { GameView, PlayerView } from "../../../core/game/GameView";
-import { renderNumber, translateText } from "../../Utils";
+import { renderNumber } from "../../../core/Util";
+import { translateText } from "../../Utils";
 import { Layer } from "./Layer";
 
 interface TeamEntry {
