+++ src/core/Util.ts
@@ -287,3 +287,30 @@ export const flattenedEmojiTable: string[] = emojiTable.flat();
 export function replacer(_key: string, value: any): any {
   return typeof value === "bigint" ? value.toString() : value;
 }
+
+export function renderTroops(troops: number): string {
+  return renderNumber(troops / 10);
+}
+
+export function renderNumber(num: number | bigint): string {
+  num = Number(num);
+  num = Math.max(num, 0);
+
+  if (num >= 10_000_000) {
+    const value = Math.floor(num / 100000) / 10;
+    return value.toFixed(1) + "M";
+  } else if (num >= 1_000_000) {
+    const value = Math.floor(num / 10000) / 100;
+    return value.toFixed(2) + "M";
+  } else if (num >= 100000) {
+    return Math.floor(num / 1000) + "K";
+  } else if (num >= 10000) {
+    const value = Math.floor(num / 100) / 10;
+    return value.toFixed(1) + "K";
+  } else if (num >= 1000) {
+    const value = Math.floor(num / 10) / 100;
+    return value.toFixed(2) + "K";
+  } else {
+    return Math.floor(num).toString();
+  }
+}
